rename sync package to externalsync
