+++ services/blockstorage/externalsync/block_sync.go
@@ -1,4 +1,4 @@
-package sync
+package externalsync
 
 import (
 	"context"
@@ -94,7 +94,7 @@ func newBlockSyncWithFactory(ctx context.Context, factory *stateFactory, config 
 	return bs
 }
 
-func NewBlockSync(ctx context.Context, config blockSyncConfig, gossip gossiptopics.BlockSync, storage BlockSyncStorage, logger log.BasicLogger, metricFactory metric.Factory) *BlockSync {
+func NewExtBlockSync(ctx context.Context, config blockSyncConfig, gossip gossiptopics.BlockSync, storage BlockSyncStorage, logger log.BasicLogger, metricFactory metric.Factory) *BlockSync {
 	conduit := &blockSyncConduit{
 		idleReset: make(chan struct{}),
 		responses: make(chan *gossipmessages.BlockAvailabilityResponseMessage),
