+++ services/blockstorage/externalsync/block_sync_gossip_client.go
@@ -1,4 +1,4 @@
-package sync
+package externalsync
 
 import (
 	"context"
