+++ services/blockstorage/externalsync/block_sync_test.go
@@ -1,4 +1,4 @@
-package sync
+package externalsync
 
 import (
 	"context"
