+++ services/blockstorage/externalsync/harness.go
@@ -1,4 +1,4 @@
-package sync
+package externalsync
 
 import (
 	"context"
