+++ services/blockstorage/externalsync/state_collecting_availability_requests.go
@@ -1,4 +1,4 @@
-package sync
+package externalsync
 
 import (
 	"context"
