+++ services/blockstorage/externalsync/state_collecting_availability_requests_test.go
@@ -1,4 +1,4 @@
-package sync
+package externalsync
 
 import (
 	"context"
