+++ services/blockstorage/externalsync/state_finished_car.go
@@ -1,4 +1,4 @@
-package sync
+package externalsync
 
 import (
 	"context"
