+++ services/blockstorage/externalsync/state_idle.go
@@ -1,4 +1,4 @@
-package sync
+package externalsync
 
 import (
 	"context"
