+++ services/blockstorage/externalsync/state_processing_blocks.go
@@ -1,4 +1,4 @@
-package sync
+package externalsync
 
 import (
 	"context"
