+++ services/blockstorage/externalsync/state_waiting_for_chunks_test.go
@@ -1,4 +1,4 @@
-package sync
+package externalsync
 
 import (
 	"context"
